Twilio courtbook refactor
